+++ popup/popup.js
@@ -7,5 +7,4 @@ document.addEventListener("DOMContentLoaded", function () {
         chrome.runtime.openOptionsPage();
         window.close();
     });
-
 });